Split ShaderGlobals definition into separate shaderglobals.h
